custom-elements/registering-custom-elements/unresolved-elements-namespace-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>If prototype interface inherits from HTMLElement then namespace must be HTML Namespace</title>
<meta name="author" title="Vasiliy Degtyarev" href="mailto:vasya@unipro.ru">
<meta name="assert" content="If prototype's interface inherits from HTMLElement, then NAMESPACE must be HTML Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var proto = Object.create(HTMLElement.prototype);
    var GeneratedConstructor = doc.registerElement('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    
    assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 'Wrong namespace');
    
}, 'Namespace should be HTML for HTMLElement');


test(function() {
    var doc = newHTMLDocument();
    
    HTML5_ELEMENTS.forEach(function(element) {
        var obj = doc.createElement(element);        
        var name = 'x-2-' + element;
        var proto = Object.create(obj.constructor.prototype);
        doc.registerElement(name, {
            prototype: Object.create(proto),
            extends: element
        });
        var customElement = doc.createElement(element, name);

        assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 
                'Wrong custom element namespace for element extending ' + element);        
    });
}, 'Namespace should be HTML for HTMLElement. Test custom element extending HTML element');

</script>
</body>
</html>
